fix(remote): avoid no-op reload audits
